% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files_from_doi.R
\name{download_files_from_doi}
\alias{download_files_from_doi}
\title{Downloads DOI files from Zenodo}
\usage{
download_files_from_doi(doi = NULL, dest_dir = ".")
}
\arguments{
\item{doi}{The DOI (Digital Object Identifier) of the deposit to be downloaded.}

\item{dest_dir}{The directory where the file(s) will be downloaded.}
}
\value{
invisibly \code{NULL}
}
\description{
This function downloads the deposits from ther DOI files from Zenodo.
}
\examples{
\dontrun{
download_files_from_doi("10.1234/abcd")
}
}
