% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_community_deposits.R
\name{get_metadata_community_deposits}
\alias{get_metadata_community_deposits}
\title{Downloads informations about the deposits of a Zenodo Community and saves them as json and parquet files}
\usage{
get_metadata_community_deposits(
  community = "ipbes",
  path = paste0(community, "_deposits")
)
}
\arguments{
\item{community}{The Zenodo community to download deposits from.}

\item{path}{The directory to save the parquet files to.}
}
\value{
The fully qualified file names of the files created:
\itemize{
\item \code{records.json}: the json file as received upon the first request
\item \code{hits.parquet}: the parquet file containing all the individual deposits
\item \code{metadata.parquet}: the parquet file containing the metadata for all the individual deposits
\item \code{related.parquet}: the parquet file containing the related identifiers for all the individual deposits including their relatinship
}
}
\description{
This function downloads deposits from a Zenodo community and saves them as parquet files.
Only published deposits are downloaded.
}
